market-discrepancy-scanner/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.54
requests>=2.32.0
python-dotenv>=1.0.1
scikit-learn>=1.5.0
joblib>=1.4.0
pytest>=8.2.0
